refactor: Move token holdings logic from BankView to TokenHoldingsSection
